Fixed: SC simulation cannot use " for aron compilation with gcc, fixed compiler warning in sc.h

+++ plugins/de.cau.cs.kieler.sc/sc/sc.h
@@ -297,7 +297,7 @@ unsigned int _setPart;
 
 #if _SC_ID_MAX < _setPartSize          // Is one _setPartType big enough?
 # define selectCid()   _SC_ERROR_DETECT_NONE_ACTIVE \
-  _BitScanReverse(active, _cid)
+  _BitScanReverse((long) active, _cid)
 #else
 # define selectCid() {				\
     for (_i = _idsetSize - 1; ; _i--) {		\
